refactor: move auth-provider under providers folder (#19)
